无尽工作台性能优化

+++ src/main/java/com/yuo/endless/Items/Tool/InfinityArrow.java
@@ -27,7 +27,7 @@ public class InfinityArrow extends Item  {
         super(new Properties().group(ModGroup.endless).maxStackSize(1).maxDamage(-1));
         ImmutableMultimap.Builder<Attribute, AttributeModifier> builder = ImmutableMultimap.builder();
         builder.put(Attributes.ATTACK_DAMAGE, new AttributeModifier(ATTACK_DAMAGE_MODIFIER, "Tool modifier", 98.0D, AttributeModifier.Operation.ADDITION));
-        builder.put(Attributes.ATTACK_SPEED, new AttributeModifier(ATTACK_SPEED_MODIFIER, "Tool modifier", (double)-2.5F, AttributeModifier.Operation.ADDITION));
+        builder.put(Attributes.ATTACK_SPEED, new AttributeModifier(ATTACK_SPEED_MODIFIER, "Tool modifier", -2.5F, AttributeModifier.Operation.ADDITION));
         this.tridentAttributes = builder.build();
     }
